+++ .github/workflows/auto-generator.yml
@@ -4,7 +4,9 @@ on:
   workflow_dispatch:
   push:
     paths:
-      - ".github/md-generator/md-generator.py"
+      - ".github/md-generator/generate_markdown.py"
+      - ".github/md-generator/generate_icon_table.py"
+      - ".github/md-generator/generate_graph.py"
       - ".github/workflows/auto-generator.yml"
       - "icons/"
 
@@ -18,7 +20,7 @@ jobs:
       - name: Get branch name
         uses: rlespinasse/github-slug-action@v3.x
 
-      - run: sudo python3 .github/md-generator/md-generator.py icons
+      - run: sudo python3 .github/md-generator/generate_markdown.py icons
 
       - name: Commit & Push in ${{ env.GITHUB_REF_SLUG_URL }}
         run: |
